refactor(lsp): extract selection range provider into perl-lsp-selection-range microcrate
Move the self-contained textDocument/selectionRange handler (636 lines)
from perl-lsp/src/features/lsp_selection_range.rs into its own
perl-lsp-selection-range crate. The original file becomes a thin
re-export. Only external dependency is lsp-types.
