+++ Cargo.toml
@@ -61,6 +61,7 @@ members = [
     "crates/perl-lsp-completion-item",
     "crates/perl-lsp-code-actions",
     "crates/perl-lsp-folding",
+    "crates/perl-lsp-selection-range",
     "crates/perl-lsp-ast-utils",
     "crates/perl-lsp-import-management",
     "crates/perl-lsp-text-utils",
@@ -189,6 +190,7 @@ allow = [
     "perl-lsp-rename",
     "perl-lsp-code-actions",
     "perl-lsp-folding",
+    "perl-lsp-selection-range",
     "perl-lsp-completion",
     "perl-lsp-file-completion",
     "perl-lsp-completion-item",
@@ -404,6 +406,7 @@ perl-lsp-file-completion = { path = "crates/perl-lsp-file-completion", version =
 perl-lsp-completion-item = { path = "crates/perl-lsp-completion-item", version = "0.11.0" }
 perl-lsp-code-actions = { path = "crates/perl-lsp-code-actions", version = "0.11.0" }
 perl-lsp-folding = { path = "crates/perl-lsp-folding", version = "0.11.0" }
+perl-lsp-selection-range = { path = "crates/perl-lsp-selection-range", version = "0.11.0" }
 perl-diagnostics-codes = { path = "crates/perl-diagnostics-codes", version = "0.11.0" }
 
 # Tier 3: Two-level dependencies
