+++ crates/perl-lsp-selection-range/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "perl-lsp-selection-range"
+version = "0.11.0"
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "SRP microcrate for Perl LSP smart selection range expansion"
+documentation = "https://docs.rs/perl-lsp-selection-range"
+keywords = ["perl", "lsp", "selection-range", "ide", "language-server"]
+categories = ["development-tools", "text-editors"]
+include = [
+  "src/**",
+  "Cargo.toml",
+  "LICENSE*",
+  "README.md"
+]
+
+[dependencies]
+lsp-types = "0.97.0"
+
+[lints]
+workspace = true
