+++ crates/perl-lsp/Cargo.toml
@@ -43,6 +43,7 @@ perl-lsp-code-lens = { workspace = true }
 perl-lsp-navigation = { workspace = true, features = ["lsp-compat"] }
 perl-lsp-type-hierarchy = { workspace = true }
 perl-lsp-rename = { workspace = true }
+perl-lsp-selection-range = { workspace = true }
 perl-lsp-semantic-tokens = { workspace = true }
 perl-lsp-launcher = { workspace = true }
 perl-lsp-diagnostic-catalog = { workspace = true }
